requirements.txt
absl-py==0.7.0
astor==0.7.1
Click==7.0
dlib==19.16.0
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.16.0
six==1.12.0
termcolor==1.1.0
tqdm==4.30.0
Werkzeug==0.14.1
paho-mqtt==1.4.0

